python_app/requirements.txt
flask==2.0.1
werkzeug==2.0.1
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
opencv-python==4.8.1.78
numpy==1.24.3
tensorflow==2.13.0
keras==2.13.1
h5py==3.9.0 